consumer/requirements.txt
kafka-python
questdb